Beautified login page on front end
