--- frontend/src/components/login.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { useAuth0 } from "@auth0/auth0-react";
-
-const LoginButton = (event) => {
-  const { loginWithRedirect } = useAuth0();
-
-  return <button onClick={() => loginWithRedirect()}>Log In</button>;
-};
-
-export default LoginButton;
