+++ frontend/src/components/loginButton.css
@@ -1,0 +1,10 @@
+.loginButton {
+  background-color: transparent;
+  border: transparent;
+  cursor: pointer;
+  height: 200px;
+  width: 550px;
+  position: relative;
+  left: 500px;
+  top: 400px;
+}
