+++ frontend/src/components/loginPage.css
@@ -1,0 +1,28 @@
+.mushroom {
+  position: relative;
+  left: 1050px;
+  top: 180px;
+  height: 180px;
+}
+
+.music {
+  position: relative;
+  top: 500px;
+  left: 200px;
+  /* top: 600px;
+  right: 250px; */
+}
+
+.header {
+  background-image: url("./background2.png");
+  background-repeat: no-repeat;
+  background-size: cover;
+  height: 790px;
+  /* display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white; */
+  cursor: url("https://cur.cursors-4u.net/games/gam-4/gam308.cur"), default !important;
+}
