+++ frontend/src/components/logoutButton.css
@@ -1,0 +1,12 @@
+.logoutButton {
+  background-color: transparent;
+  border: transparent;
+  cursor: pointer;
+  height: 100px;
+  width: 520px;
+  position: relative;
+  right: 30px;
+  top: 520px;
+  /* left: 750px;
+  top: 520px; */
+}
